+++ flutter_sound/lib/public/flutter_sound_recorder.dart
@@ -64,17 +64,15 @@ class FlutterSoundRecorder implements FlutterSoundRecorderCallback {
 
   /// Used if the App wants to dynamically change the Log Level.
   /// Seldom used. Most of the time the Log Level is specified during the constructor.
-  Future<void> setLogLevel(Level aLevel) async {
+  void setLogLevel(Level aLevel) {
     _logLevel = aLevel;
     _logger = Logger(level: aLevel);
-    await _lock.synchronized(() async {
-      if (_isInited != Initialized.notInitialized) {
-        await FlutterSoundRecorderPlatform.instance.setLogLevel(
-          this,
-          aLevel,
-        );
-      }
-    });
+    if (_isInited != Initialized.notInitialized) {
+      FlutterSoundRecorderPlatform.instance.setLogLevel(
+        this,
+        aLevel,
+      );
+    }
   }
 
   /// Locals